+++ rust/candid/src/bindings/typescript.rs
@@ -78,26 +78,22 @@ fn pp_field<'a>(env: &'a TypeEnv, field: &'a Field, is_ref: bool) -> RcDoc<'a> {
 }
 
 fn pp_function<'a>(env: &'a TypeEnv, func: &'a Function) -> RcDoc<'a> {
-    let args = func
-        .args
-        .iter()
-        .enumerate()
-        .map(|(i, ty)| RcDoc::text(format!("arg_{}: ", i)).append(pp_ty(env, ty, true)));
-    let args = enclose("(", concat(args, ","), ")");
-    let rets = str("Promise").append(enclose(
-        "<",
-        match func.rets.len() {
-            0 => str("undefined"),
-            1 => pp_ty(env, &func.rets[0], true),
-            _ => enclose(
-                "[",
-                concat(func.rets.iter().map(|ty| pp_ty(env, ty, true)), ","),
-                "]",
-            ),
-        },
+    let args = func.args.iter().map(|ty| pp_ty(env, ty, true));
+    let args = enclose("[", concat(args, ","), "]");
+    let rets = match func.rets.len() {
+        0 => str("undefined"),
+        1 => pp_ty(env, &func.rets[0], true),
+        _ => enclose(
+            "[",
+            concat(func.rets.iter().map(|ty| pp_ty(env, ty, true)), ","),
+            "]",
+        ),
+    };
+    enclose(
+        "ActorMethod<",
+        concat([args, rets].iter().cloned(), ","),
         ">",
-    ));
-    args.append(" => ").append(rets).nest(INDENT_SPACE)
+    )
 }
 
 fn pp_service<'a>(env: &'a TypeEnv, serv: &'a [(String, Type)]) -> RcDoc<'a> {
@@ -153,7 +149,9 @@ fn pp_actor<'a>(env: &'a TypeEnv, ty: &'a Type) -> RcDoc<'a> {
 }
 
 pub fn compile(env: &TypeEnv, actor: &Option<Type>) -> String {
-    let header = r#"import type { Principal } from '@dfinity/principal';"#;
+    let header = r#"import type { Principal } from '@dfinity/principal';
+import type { ActorMethod } from '@dfinity/agent';
+"#;
     let def_list: Vec<_> = env.0.iter().map(|pair| pair.0.as_ref()).collect();
     let defs = pp_defs(env, &def_list);
     let actor = match actor {