+++ rust/candid/tests/assets/ok/keyword.d.ts
@@ -1,4 +1,6 @@
 import type { Principal } from '@dfinity/principal';
+import type { ActorMethod } from '@dfinity/agent';
+
 export type if_ = {
     'branch' : { 'val' : bigint, 'left' : if_, 'right' : if_ }
   } |
@@ -6,29 +8,22 @@ export type if_ = {
 export type list = [] | [node];
 export interface node { 'head' : bigint, 'tail' : list }
 export type o = [] | [o];
-export interface return_ {
-  'f' : t,
-  'g' : (arg_0: list) => Promise<[if_, stream]>,
-}
+export interface return_ { 'f' : t, 'g' : ActorMethod<[list], [if_, stream]> }
 export type stream = [] | [{ 'head' : bigint, 'next' : [Principal, string] }];
-export type t = (arg_0: Principal) => Promise<undefined>;
+export type t = ActorMethod<[Principal], undefined>;
 export interface _SERVICE {
-  'Oneway' : () => Promise<undefined>,
-  'f_' : (arg_0: o) => Promise<o>,
-  'field' : (arg_0: { 'test' : number, _1291438163_ : number }) => Promise<{}>,
-  'fieldnat' : (arg_0: { _2_ : bigint, '2' : bigint }) => Promise<[bigint]>,
-  'oneway' : (arg_0: number) => Promise<undefined>,
-  'oneway_' : (arg_0: number) => Promise<undefined>,
-  'query' : (arg_0: Array<number>) => Promise<Array<number>>,
-  'return' : (arg_0: o) => Promise<o>,
+  'Oneway' : ActorMethod<[], undefined>,
+  'f_' : ActorMethod<[o], o>,
+  'field' : ActorMethod<[{ 'test' : number, _1291438163_ : number }], {}>,
+  'fieldnat' : ActorMethod<[{ _2_ : bigint, '2' : bigint }], [bigint]>,
+  'oneway' : ActorMethod<[number], undefined>,
+  'oneway_' : ActorMethod<[number], undefined>,
+  'query' : ActorMethod<[Array<number>], Array<number>>,
+  'return' : ActorMethod<[o], o>,
   'service' : t,
-  'tuple' : (arg_0: [bigint, Array<number>, string]) => Promise<
-      [bigint, number]
-    >,
-  'variant' : (
-      arg_0: { 'A' : null } |
-        { 'B' : null } |
-        { 'C' : null } |
-        { 'D' : number },
-    ) => Promise<undefined>,
+  'tuple' : ActorMethod<[[bigint, Array<number>, string]], [bigint, number]>,
+  'variant' : ActorMethod<
+    [{ 'A' : null } | { 'B' : null } | { 'C' : null } | { 'D' : number }],
+    undefined,
+  >,
 }