+++ rust/candid/tests/assets/ok/service.d.ts
@@ -1,15 +1,19 @@
 import type { Principal } from '@dfinity/principal';
-export type Func = () => Promise<Principal>;
+import type { ActorMethod } from '@dfinity/agent';
+
+export type Func = ActorMethod<[], Principal>;
 export interface Service { 'f' : Func }
 export type Service2 = Service;
 export interface _SERVICE {
-  'asArray' : () => Promise<[Array<Principal>, Array<[Principal, string]>]>,
-  'asPrincipal' : () => Promise<[Principal, [Principal, string]]>,
-  'asRecord' : () => Promise<
-      [Principal, [] | [Principal], [Principal, string]]
-    >,
-  'asVariant' : () => Promise<
-      { 'a' : Principal } |
-        { 'b' : { 'f' : [] | [[Principal, string]] } }
-    >,
+  'asArray' : ActorMethod<[], [Array<Principal>, Array<[Principal, string]>]>,
+  'asPrincipal' : ActorMethod<[], [Principal, [Principal, string]]>,
+  'asRecord' : ActorMethod<
+    [],
+    [Principal, [] | [Principal], [Principal, string]],
+  >,
+  'asVariant' : ActorMethod<
+    [],
+    { 'a' : Principal } |
+      { 'b' : { 'f' : [] | [[Principal, string]] } },
+  >,
 }